refactor(public_api): move token-create logic into serializer, bulk-create grants
Drive POST /public-api-tokens/ through SystemUserTokenCreateSerializer.create()
instead of the view: membership resolution, SystemUser provisioning via the
injected auth service, and ResourceAccess grants now live in the serializer.
The view is reduced to validate -> save -> serialize response.

Grants are now persisted with a single bulk_create instead of one INSERT per
resource. Duplicate integration_name returns a DRF-standard field error via
serializers.ValidationError; no-membership raises PermissionDenied (403).
