+++ public_api/views.py
@@ -1,9 +1,5 @@
 import logging
-from typing import Annotated
 
-from django.db import IntegrityError, transaction
-
-from dependency_injector.wiring import Provide, inject
 from drf_spectacular.utils import extend_schema
 from rest_framework import status
 from rest_framework.request import Request
@@ -12,12 +8,11 @@ from rest_framework.viewsets import GenericViewSet
 
 from organizations.models import get_active_organization_membership
 from organizations.permissions import IsOrganizationAdmin
-from public_api.models import ResourceAccess, SystemUser
+from public_api.models import SystemUser
 from public_api.serializers import (
     SystemUserTokenCreateSerializer,
     SystemUserTokenResponseSerializer,
 )
-from public_api.services import PublicAPIAuthService
 
 
 logger = logging.getLogger(__name__)
@@ -35,18 +30,6 @@ class SystemUserTokenViewSet(GenericViewSet):
 
     permission_classes = (IsOrganizationAdmin,)
     serializer_class = SystemUserTokenCreateSerializer
-
-    @inject
-    def __init__(
-        self,
-        *args,
-        public_api_auth_service: Annotated[
-            PublicAPIAuthService, Provide["public_api_auth_service"]
-        ],
-        **kwargs,
-    ) -> None:
-        super().__init__(*args, **kwargs)
-        self.public_api_auth_service = public_api_auth_service
 
     def get_queryset(self):  # type: ignore[override]
         """Org-scoped queryset — used for router introspection."""
@@ -76,43 +59,10 @@ class SystemUserTokenViewSet(GenericViewSet):
 
         HTTP 403 is returned for non-admin callers; HTTP 401 for unauthenticated.
         """
-        membership = get_active_organization_membership(request.user)  # type: ignore[arg-type]
-        if membership is None:
-            # IsOrganizationAdmin.has_permission already guards this; defensive fallback.
-            return Response(
-                {"detail": "No active organisation membership."},
-                status=status.HTTP_403_FORBIDDEN,
-            )
-
-        serializer = SystemUserTokenCreateSerializer(data=request.data)
+        serializer = self.get_serializer(data=request.data)
         serializer.is_valid(raise_exception=True)
+        system_user = serializer.save()
 
-        integration_name: str = serializer.validated_data["integration_name"]
-        available_resources: list[str] = serializer.validated_data["available_resources"]
-        organization = membership.organization
-
-        try:
-            with transaction.atomic():
-                system_user, plaintext_token = self.public_api_auth_service.create_system_user(
-                    integration_name=integration_name,
-                    organization=organization,
-                )
-        except IntegrityError:
-            return Response(
-                {
-                    "integration_name": [
-                        f"A token with integration_name '{integration_name}' already exists."
-                    ]
-                },
-                status=status.HTTP_400_BAD_REQUEST,
-            )
-
-        # Persist one ResourceAccess row per requested resource.
-        for resource_name in available_resources:
-            ResourceAccess.objects.create(system_user=system_user, resource_name=resource_name)
-
-        # Build the response data: include the plaintext token ONCE via a pseudo-attribute.
-        system_user.token = plaintext_token  # type: ignore[attr-defined]
         response_serializer = SystemUserTokenResponseSerializer(
             system_user, context={"request": request}
         )
