+++ schema.yml
@@ -8647,8 +8647,9 @@ components:
         Input serializer for creating a new public-API token (SystemUser + ResourceAccess rows).
 
         Accepts ``integration_name`` and ``available_resources`` (a non-empty list of valid
-        ``PublicAPIResources`` values).  On a successful create the view adds a write-once
-        ``token`` field to the response data; this serializer never stores or exposes
+        ``PublicAPIResources`` values).  ``create()`` provisions the ``SystemUser`` via the
+        injected auth service and bulk-creates the ``ResourceAccess`` grants, attaching the
+        write-once plaintext ``token`` to the returned instance.  Never stores or exposes
         ``long_lived_token_hash``.
       properties:
         integration_name:
